Began day 12
e  need a Recursive function with input as an Adjaceny list. Bunch  of  edges.

We then do a Depth first  search.

While data is ingested build a dict of  sets. Each set  shows an edge.

The first visited edges are stored in a stack. LIFO data structure.

The goal is to find the number of distinct paths that start at start, end at end.

Need to keep track of  the small caves.

When the End cave is reached , increment  the  number  of  paths.

Backtracking needed to remove small caves. To avoid processing a node more than once, use a boolean visited array.